requirements.txt
absl-py==2.1.0
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aiosignal==1.3.2
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
array_record==0.6.0
astunparse==1.6.3
attrs==25.1.0
braceexpand==0.1.7
certifi==2024.8.30
cfgv==3.4.0
charset-normalizer==3.3.2
chex==0.1.86
click==8.1.8
cloudpickle==3.1.1
compoconf==0.1.0
contextlib2==21.6.0
contourpy==1.3.0
cycler==0.12.1
dacite==1.8.1
datasets==3.3.1
dill==0.3.8
distlib==0.3.8
dm-tree==0.1.9
dm_pix==0.4.4
docker-pycreds==0.4.0
docstring_parser==0.16
editables==0.5
einops==0.8.1
etils==1.9.4
filelock==3.15.4
flatbuffers==24.3.25
flax==0.10.2
fonttools==4.53.1
frozenlist==1.5.0
fsspec==2024.9.0
gast==0.6.0
gitdb==4.0.12
GitPython==3.1.44
google-pasta==0.2.0
googleapis-common-protos==1.67.0
grain==0.2.3
grpcio==1.66.1
gviz-api==1.10.0
h5py==3.11.0
hatchling==1.27.0
huggingface-hub==0.28.1
humanize==4.10.0
hydra-core==1.3.2
identify==2.6.0
idna==3.8
immutabledict==4.2.1
importlib_resources==6.4.4
iniconfig==2.0.0
jax==0.4.32
jax-cuda12-pjrt==0.4.32
jax-cuda12-plugin==0.4.32
jaxlib==0.4.32
jaxtyping==0.2.38
Jinja2==3.1.3
keras==3.5.0
kiwisolver==1.4.7
libclang==18.1.1
lightning-utilities==0.12.0
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.2
mdurl==0.1.2
ml-dtypes==0.4.0
ml_collections==0.1.1
more-itertools==10.6.0
mpmath==1.3.0
msgpack==1.0.8
multidict==6.1.0
multiprocess==0.70.16
namex==0.0.8
nest-asyncio==1.6.0
networkx==3.2.1
nodeenv==1.9.1
numpy==1.26.4
# nvidia-cublas-cu12==12.6.1.4
# nvidia-cuda-cupti-cu12==12.6.68
# nvidia-cuda-nvcc-cu12==12.6.68
# nvidia-cuda-runtime-cu12==12.6.68
# nvidia-cudnn-cu12==9.4.0.58
# nvidia-cufft-cu12==11.2.6.59
# nvidia-cusolver-cu12==11.6.4.69
# nvidia-cusparse-cu12==12.5.3.3
# nvidia-nccl-cu12==2.22.3
# nvidia-nvjitlink-cu12==12.6.68
omegaconf==2.3.0
opt-einsum==3.3.0
optax==0.2.3
optree==0.12.1
orbax==0.1.9
orbax-checkpoint==0.6.1
packaging==24.2
pandas==2.2.2
pathspec==0.12.1
pillow==10.4.0
platformdirs==4.2.2
pluggy==1.5.0
pre-commit==3.8.0
promise==2.3
propcache==0.2.1
# protobuf==4.25.4
psutil==7.0.0
pyarrow==19.0.0
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.18.0
pyparsing==3.1.4
pytest==8.3.2
python-dateutil==2.9.0.post0
pytorch-lightning==2.5.0.post0
pytz==2024.1
PyYAML==6.0.2
requests==2.32.3
rich==13.8.0
safetensors==0.5.2
scipy==1.14.1
seaborn==0.13.2
sentry-sdk==2.22.0
setproctitle==1.3.4
simple-parsing==0.1.7
six==1.16.0
smmap==5.0.2
sympy==1.12
tabulate==0.9.0
tensorboard==2.17.1
tensorboard-data-server==0.7.2
tensorboard-plugin-profile==2.17.0
tensorflow==2.17.0
tensorflow-datasets==4.9.7
tensorflow-io-gcs-filesystem==0.37.1
tensorflow-metadata==1.16.1
tensorstore==0.1.65
termcolor==2.4.0
timm==1.0.14
toml==0.10.2
toolz==0.12.1
# torch==2.4.1
torch
# torchaudio==2.4.1
torchaudio
torchmetrics==1.6.1
# torchvision==0.19.1
torchvision
tqdm==4.66.5
trove-classifiers==2025.1.15.22
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
virtualenv==20.26.3
wadler_lindig==0.1.3
wandb==0.19.6
webdataset==0.2.111
Werkzeug==3.0.4
wrapt==1.16.0
xxhash==3.5.0
yarl==1.18.3
zipp==3.20.1
